feat(#DBSYNC-43): Dropbox-style tray flyout + separate setup window
Replace the 760x540 dashboard with a compact borderless tray flyout and a
separate decorated setup window.

Rust (S1): two static windows in tauri.conf (main=flyout borderless/
skipTaskbar/resizable:false/360x600/alwaysOnTop, setup=decorated 640x520),
capability grants 'setup'. Tray single Click toggles+positions the flyout
(DPI-scaled from the tray rect, primary+bottom-taskbar, clamped);
DoubleClick + 'Open Dashboard' menu item removed. Hide-on-blur guarded by a
~250ms LAST_TOGGLE stamp so tray-click-while-open closes cleanly. run_events
hides main on blur/close but genuinely closes setup. setup() shows the setup
window when onboarding is needed; flyout appears only on tray click. New
show_setup_window command.

Frontend (S2+S3, absorbs DBSYNC-20): App.tsx (883 lines) deleted, split into
7 focused hooks (useSyncDashboard/useActivityLog/useAuthFlow/useStartup
Requirements/useCloudscPlaceholders/useRemoteBrowser/useSelectiveSync) + two
window roots routed by getCurrentWindow().label: FlyoutApp (Accueil = recent
+ activity, Dossiers = READ-ONLY browse, Activite = full log/conflicts, status
footer) and SetupApp (connect/folder/re-auth, self-closes when configured).
All in-app hydrate/download buttons removed (hydration is double-click on the
.cloudsc placeholder). The startup show/hide useEffect removed (Rust owns it).
DBSYNC-21 try/catch + DBSYNC-22 oauthCompletedRef guard preserved. Declared
@dropbox-sync/shared as a dependency.

#DBSYNC-43
