+++ apps/desktop/src-tauri/src/run_events.rs
@@ -1,4 +1,5 @@
 use std::path::Path;
+use std::time::Duration;
 
 use tauri::AppHandle;
 use tauri::DragDropEvent;
@@ -36,10 +37,28 @@ pub(crate) fn handle_run_event(app_handle: &AppHandle, event: RunEvent) {
         }
         RunEvent::WindowEvent { label, event: win_evt, .. } => {
             match win_evt {
+                // `main` is a tray flyout, not the app: closing it (e.g. Alt+F4) just
+                // hides it, it never quits the app.
                 WindowEvent::CloseRequested { api, .. } if label == "main" => {
                     api.prevent_close();
                     if let Some(w) = app_handle.get_webview_window("main") {
                         let _ = w.hide();
+                    }
+                }
+                // `setup` is a real, closable window (not a flyout) — let it close
+                // genuinely; do not prevent-close/hide it like `main`.
+                WindowEvent::CloseRequested { .. } if label == "setup" => {}
+                // Hide the flyout when it loses focus (click-away), unless it was *just*
+                // toggled from the tray — otherwise a tray click on an already-focused
+                // flyout would blur-hide it and then immediately re-show it, or a click
+                // that both blurs a previous window and shows the flyout could race the
+                // blur-hide against the tray's own show.
+                WindowEvent::Focused(false) if label == "main" => {
+                    if let Some(w) = app_handle.get_webview_window("main") {
+                        let visible = w.is_visible().unwrap_or(false);
+                        if visible && !crate::recently_toggled(Duration::from_millis(250)) {
+                            let _ = w.hide();
+                        }
                     }
                 }
                 WindowEvent::DragDrop(DragDropEvent::Drop { paths, .. }) => {
