+++ apps/desktop/src/components/FlyoutApp.css
@@ -1,0 +1,282 @@
+:root {
+  --flyout-bg: #f7f9fc;
+  --flyout-panel: #ffffff;
+  --flyout-border: #d9e3f2;
+  --flyout-text: #0f172a;
+  --flyout-muted: #64748b;
+  --flyout-accent: #2563eb;
+  --flyout-accent-soft: rgba(37, 99, 235, 0.12);
+  --flyout-error: #c0392b;
+  --flyout-warn: #b45309;
+  --flyout-ok: #16a34a;
+}
+
+@media (prefers-color-scheme: dark) {
+  :root {
+    --flyout-bg: #111827;
+    --flyout-panel: #1a2233;
+    --flyout-border: #2a3550;
+    --flyout-text: #e5e9f2;
+    --flyout-muted: #94a3b8;
+    --flyout-accent: #60a5fa;
+    --flyout-accent-soft: rgba(96, 165, 250, 0.18);
+    --flyout-error: #f87171;
+    --flyout-warn: #fbbf24;
+    --flyout-ok: #4ade80;
+  }
+}
+
+html, body, #root {
+  height: 100%;
+  margin: 0;
+}
+
+.flyout-btn {
+  background: linear-gradient(180deg, #3b82f6 0%, #2563eb 100%);
+  color: white;
+  border: none;
+  border-radius: 8px;
+  padding: 5px 10px;
+  font-weight: 600;
+  font-size: 11px;
+  cursor: pointer;
+  box-shadow: none;
+}
+
+.flyout-btn:disabled {
+  opacity: 0.5;
+  cursor: not-allowed;
+}
+
+.flyout {
+  display: flex;
+  flex-direction: column;
+  height: 100vh;
+  width: 100%;
+  box-sizing: border-box;
+  background: var(--flyout-bg);
+  color: var(--flyout-text);
+  font-family: Inter, -apple-system, BlinkMacSystemFont, "Segoe UI", Arial, sans-serif;
+  font-size: 12.5px;
+  overflow: hidden;
+}
+
+.flyout-nav {
+  display: flex;
+  align-items: center;
+  justify-content: space-between;
+  gap: 4px;
+  padding: 6px 8px;
+  border-bottom: 1px solid var(--flyout-border);
+  background: var(--flyout-panel);
+  flex: 0 0 auto;
+}
+
+.flyout-nav-items {
+  display: flex;
+  gap: 2px;
+}
+
+.flyout-nav-item {
+  display: flex;
+  flex-direction: column;
+  align-items: center;
+  gap: 2px;
+  background: transparent;
+  color: var(--flyout-muted);
+  border: none;
+  border-radius: 8px;
+  padding: 4px 10px;
+  margin: 0;
+  font-weight: 500;
+  font-size: 10.5px;
+  box-shadow: none;
+  cursor: pointer;
+  transition: background 0.12s ease, color 0.12s ease;
+}
+
+.flyout-nav-item:hover:not(:disabled) {
+  background: var(--flyout-accent-soft);
+  transform: none;
+  box-shadow: none;
+  filter: none;
+}
+
+.flyout-nav-item.active {
+  background: var(--flyout-accent-soft);
+  color: var(--flyout-accent);
+}
+
+.flyout-nav-item .glyph {
+  font-size: 15px;
+  line-height: 1;
+}
+
+.flyout-settings {
+  background: transparent;
+  border: none;
+  box-shadow: none;
+  cursor: pointer;
+  padding: 4px 8px;
+  margin: 0;
+  font-size: 14px;
+  color: var(--flyout-muted);
+}
+
+.flyout-settings:hover {
+  color: var(--flyout-accent);
+  transform: none;
+  filter: none;
+  box-shadow: none;
+}
+
+.flyout-body {
+  flex: 1 1 auto;
+  overflow-y: auto;
+  padding: 10px 12px;
+}
+
+.flyout-banner {
+  display: flex;
+  align-items: center;
+  justify-content: space-between;
+  gap: 8px;
+  background: var(--flyout-accent-soft);
+  border: 1px solid var(--flyout-border);
+  border-radius: 8px;
+  padding: 8px 10px;
+  margin-bottom: 10px;
+}
+
+.flyout-banner p {
+  margin: 0;
+  color: var(--flyout-text);
+  font-size: 12px;
+}
+
+.flyout-section h2 {
+  font-size: 11px;
+  text-transform: uppercase;
+  letter-spacing: 0.05em;
+  color: var(--flyout-muted);
+  margin: 14px 0 6px;
+}
+
+.flyout-section h2:first-child {
+  margin-top: 0;
+}
+
+.flyout-path-bar {
+  display: flex;
+  align-items: center;
+  gap: 8px;
+  margin-bottom: 6px;
+}
+
+.flyout-path-bar .path {
+  font-size: 11.5px;
+  color: var(--flyout-muted);
+  word-break: break-all;
+}
+
+.flyout-retry {
+  font-size: 11px;
+  padding: 4px 10px;
+  margin-bottom: 6px;
+}
+
+.flyout-list {
+  list-style: none;
+  margin: 0;
+  padding: 0;
+  display: flex;
+  flex-direction: column;
+  gap: 4px;
+}
+
+.flyout-list li {
+  background: var(--flyout-panel);
+  border: 1px solid var(--flyout-border);
+  border-radius: 8px;
+  padding: 6px 8px;
+  margin: 0;
+}
+
+.flyout-list.compact li {
+  padding: 5px 8px;
+}
+
+.flyout-list .muted {
+  color: var(--flyout-muted);
+  background: transparent;
+  border: none;
+  padding: 4px 2px;
+}
+
+.flyout-list .row-link {
+  display: block;
+  width: 100%;
+  text-align: left;
+  background: transparent;
+  border: none;
+  box-shadow: none;
+  padding: 0;
+  margin: 0;
+  color: inherit;
+  cursor: pointer;
+  font: inherit;
+}
+
+.flyout-list .row-link:hover {
+  transform: none;
+  filter: none;
+  box-shadow: none;
+  color: var(--flyout-accent);
+}
+
+.row-main {
+  font-size: 12px;
+  word-break: break-word;
+}
+
+.row-sub {
+  font-size: 10.5px;
+  color: var(--flyout-muted);
+  margin-top: 2px;
+}
+
+.row-sub.error {
+  color: var(--flyout-error);
+}
+
+.flyout-footer {
+  flex: 0 0 auto;
+  display: flex;
+  align-items: center;
+  gap: 6px;
+  padding: 7px 12px;
+  font-size: 11px;
+  border-top: 1px solid var(--flyout-border);
+  background: var(--flyout-panel);
+  color: var(--flyout-muted);
+}
+
+.flyout-footer .dot {
+  width: 7px;
+  height: 7px;
+  border-radius: 50%;
+  background: var(--flyout-ok);
+  flex: 0 0 auto;
+}
+
+.flyout-footer.busy .dot {
+  background: var(--flyout-accent);
+}
+
+.flyout-footer.warn .dot {
+  background: var(--flyout-warn);
+}
+
+.flyout-footer.ok .dot {
+  background: var(--flyout-ok);
+}
